+++ Classes/Configuration/WidgetsProvider.php
@@ -33,6 +33,7 @@ final class WidgetsProvider
         'contentNames' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.contents.contentNames.title',
         'contentPieces' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.contents.contentPieces.title',
         'countries' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.userCountry.country.title',
+        'javaScriptErrors' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.events.javaScriptErrors.title',
         'linkMatomo' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.linkMatomo.title',
         'osFamilies' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.devicesDetection.osFamilies.title',
         'siteSearchKeywords' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.actions.siteSearchKeywords.title',